move struct locations
